requirements.txt
fastapi
uvicorn
gunicorn
pillow
numpy
daltonlens
python-multipart
